Feat : Packet Service, Controller 추가

+++ src/main/java/com/ssinhwa/gameserver/redisserver/dto/PlayerContinuousData.java
@@ -13,5 +13,5 @@ public class PlayerContinuousData {
     private Integer playerId;
     private Integer playerState;
     private WPosition worldPosition;
-    private Float rotation;
+    private double rotation;
 }
